src/requirements.txt
Pillow==10.1.0
pynput==1.7.6
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
